la plateforme est jaune

+++ UI/InGame/Gestion/blocks/OneItemSliderBlock.cs
@@ -7,7 +7,7 @@ namespace UI.InGame.Gestion.blocks {
         [SerializeField] private TextMeshProUGUI sliderName;
         [SerializeField] private Slider _slider;
 
-        public void SetBlock(string sliderName, int value, int maxValue) {
+        public void SetBlock(string sliderName, float value, int maxValue) {
             this.sliderName.text = sliderName;
             _slider.maxValue = maxValue;
             _slider.value = value;
